backend/requirements.txt
flask
flask-cors
google-generativeai
pypdf
pdfplumber
faiss-cpu
numpy
python-dotenv
